Extract common code to src/common/sources.ts
This is especially a preparation for hasRestrictedAccessLevel, which
will be used also from main in the next commit.
