+++ src/common/legacySource.ts
@@ -4,8 +4,6 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-import type { Source } from '../ipc/sources';
-
 export const isLegacyUrl = (url: string) =>
     url.match(/^https?:\/\/developer\.nordicsemi\.com\//);
 
@@ -30,21 +28,3 @@ export const migrateURL = (url: string) =>
 
 export const migrateAllURLsInJSON = (json: string) =>
     json.replace(/"https?:[^"]*"/g, migrateURL);
-
-const deprecatedSources = [
-    'toolchain-manager',
-    '3.8-release-test',
-    'neutrino-external',
-    'neutrino-internal',
-    'crasher',
-    'internal',
-    'experimental',
-];
-
-const deprecatedSourceURLs = deprecatedSources.map(
-    name =>
-        `https://files.nordicsemi.com/artifactory/swtools/internal/ncd/apps/${name}/source.json`
-);
-
-export const isDeprecatedSource = (url: Source) =>
-    deprecatedSourceURLs.includes(url.url);
